[refactor] #3833, #2373, #3601: Split off Kagami (#3841)
* [refactor]: remove Swarm from Kagami; introduce refactored `iroha_swarm`


* [refactor]: remove validator from Kagami; cleanup


* [misc]: fix workspace info, introduce `wasm_builder_cli`


* [ci]: update Genesis and Swarm cmds in scripts


* [feat]: produce workable `iroha_wasm_builder_cli`


* [feat]: enhance CLI UI with spinners


* [build]: remove `iroha_wasm_builder` dep from Kagami


* [build]: tree-shake unused spinners


* [test]: fix swarm tests


* [docs]: add README for `iroha_swarm`


* [refactor]: remove `UserInterface` struct


* [feat]: add `--outfile` arg for wasm cli


* [docs]: document how to build the default validator


* [chore]: unused imports


* [docs]: link the directory to the cli


* [docs]: enhance warning about an inlined validator


* [chore]: use stdout


* [refactor]: move `wasm_builder_cli` to `./tools/`


---------
